Moved Ingredient to a models module
Sections are going to require a few so I want them organized. I may yet
still split into files.
